pyproject.toml
[tool.poetry]
name = "example"
version = "0.1.0"
description = ""
authors = ["Chris Hunt <chrahunt@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
conan = "^1.10"
requests = "^2.20"
python-daemon = "^2.2"
click = "^7.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
example = "example.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
